refactor eval runtime type helpers
